requirements/test.txt
-r base.txt
pytest==3.2.2
pytest-flask==0.10.0
pytest-mock==1.6.3